# contracts/crm-pipeline/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CRM Pipeline Service",
    "description": "CRM Pipeline: leads, opportunities, pipeline stages, activities, and sales handoff.",
    "license": { "name": "" },
    "version": "0.1.0"
  },
  "paths": {
    "/api/crm-pipeline/leads": {
      "get": {
        "tags": ["Leads"],
        "summary": "List leads",
        "operationId": "list_leads",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "assigned_user_id", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": {
          "200": { "description": "Paginated list of leads", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Lead" } } } } }
        },
        "security": [{ "bearer": [] }]
      },
      "post": {
        "tags": ["Leads"],
        "summary": "Create a lead",
        "operationId": "create_lead",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateLeadRequest" } } } },
        "responses": {
          "201": { "description": "Lead created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Lead" } } } },
          "422": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } }
        },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/leads/{id}": {
      "get": {
        "tags": ["Leads"],
        "summary": "Get a lead",
        "operationId": "get_lead",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Lead detail", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Lead" } } } },
          "404": { "description": "Not found" }
        },
        "security": [{ "bearer": [] }]
      },
      "put": {
        "tags": ["Leads"],
        "summary": "Update a lead",
        "operationId": "update_lead",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateLeadRequest" } } } },
        "responses": {
          "200": { "description": "Lead updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Lead" } } } }
        },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/leads/{id}/contact": {
      "post": {
        "tags": ["Leads"],
        "summary": "Mark lead as contacted",
        "operationId": "mark_contacted",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Lead updated" } },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/leads/{id}/qualify": {
      "post": {
        "tags": ["Leads"],
        "summary": "Mark lead as qualifying",
        "operationId": "mark_qualifying",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Lead updated" } },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/leads/{id}/mark-qualified": {
      "post": {
        "tags": ["Leads"],
        "summary": "Mark lead as qualified",
        "operationId": "mark_qualified",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Lead updated" } },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/leads/{id}/convert": {
      "post": {
        "tags": ["Leads"],
        "summary": "Convert lead to opportunity",
        "operationId": "convert_lead",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConvertLeadRequest" } } } },
        "responses": {
          "200": { "description": "Lead converted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConvertLeadResponse" } } } },
          "422": { "description": "Validation error — party_id required" }
        },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/leads/{id}/disqualify": {
      "post": {
        "tags": ["Leads"],
        "summary": "Disqualify a lead",
        "operationId": "disqualify_lead",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DisqualifyLeadRequest" } } } },
        "responses": { "200": { "description": "Lead disqualified" } },
        "security": [{ "bearer": [] }]
      }
    },
    "/api/crm-pipeline/leads/{id}/mark-dead": {
      "post": {
        "tags": ["Leads"],
        "summary": "Mark lead as dead",
        "operationId": "mark_dead",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Lead updated" } },
        "security": [{ "bearer": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Lead": {
        "type": "object",
        "required": ["id", "tenant_id", "lead_number", "title", "status", "created_at", "updated_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "tenant_id": { "type": "string" },
          "lead_number": { "type": "string" },
          "title": { "type": "string" },
          "status": { "type": "string" },
          "source": { "type": ["string", "null"] },
          "party_id": { "type": ["string", "null"], "format": "uuid" },
          "party_contact_id": { "type": ["string", "null"], "format": "uuid" },
          "assigned_user_id": { "type": ["string", "null"], "format": "uuid" },
          "estimated_value_cents": { "type": ["integer", "null"] },
          "currency_code": { "type": ["string", "null"] },
          "notes": { "type": ["string", "null"] },
          "disqualify_reason": { "type": ["string", "null"] },
          "converted_opportunity_id": { "type": ["string", "null"], "format": "uuid" },
          "converted_at": { "type": ["string", "null"], "format": "date-time" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "CreateLeadRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": { "type": "string" },
          "source": { "type": ["string", "null"] },
          "party_id": { "type": ["string", "null"], "format": "uuid" },
          "party_contact_id": { "type": ["string", "null"], "format": "uuid" },
          "assigned_user_id": { "type": ["string", "null"], "format": "uuid" },
          "estimated_value_cents": { "type": ["integer", "null"] },
          "currency_code": { "type": ["string", "null"] },
          "notes": { "type": ["string", "null"] }
        }
      },
      "UpdateLeadRequest": {
        "type": "object",
        "properties": {
          "title": { "type": ["string", "null"] },
          "source": { "type": ["string", "null"] },
          "party_id": { "type": ["string", "null"], "format": "uuid" },
          "party_contact_id": { "type": ["string", "null"], "format": "uuid" },
          "assigned_user_id": { "type": ["string", "null"], "format": "uuid" },
          "estimated_value_cents": { "type": ["integer", "null"] },
          "currency_code": { "type": ["string", "null"] },
          "notes": { "type": ["string", "null"] }
        }
      },
      "ConvertLeadRequest": {
        "type": "object",
        "required": ["party_id"],
        "properties": {
          "party_id": { "type": "string", "format": "uuid" },
          "opportunity_title": { "type": ["string", "null"] }
        }
      },
      "ConvertLeadResponse": {
        "type": "object",
        "required": ["lead_id", "opportunity_id"],
        "properties": {
          "lead_id": { "type": "string", "format": "uuid" },
          "opportunity_id": { "type": "string", "format": "uuid" }
        }
      },
      "DisqualifyLeadRequest": {
        "type": "object",
        "required": ["reason"],
        "properties": {
          "reason": { "type": "string" }
        }
      },
...
      "ApiError": {
        "type": "object",
        "required": ["error", "message"],
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" },
          "field": { "type": ["string", "null"] }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}